Ubuntu patch from 9850 for building on armel arch
git-svn-id: svn://scribus.net/branches/Version14x/Scribus@17287 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/fonts/scfontmetrics.cpp
@@ -281,7 +281,7 @@ QPixmap FontSample(const ScFace& fnt, int s, QString ts, QColor back, bool force
 			{
 				gly.translate(static_cast<qreal>(pen_x) / 6400.0, a);
 				gp = getMaxClipF(&gly);
-				ymax = qMax(ymax, gp.y());
+				ymax = qMax(double(ymax), gp.y());
 				p->setupPolygon(&gly);
 				p->fillPath();
 			}
@@ -302,7 +302,7 @@ QPixmap FontSample(const ScFace& fnt, int s, QString ts, QColor back, bool force
 			{
 				gly.translate(static_cast<qreal>(pen_x) / 6400.0, a);
 				gp = getMaxClipF(&gly);
-				ymax = qMax(ymax, gp.y());
+				ymax = qMax(double(ymax), gp.y());
 				p->setupPolygon(&gly);
 				p->fillPath();
 			}
