+++ scribus/plugins/svgexplugin/svgexplugin.cpp
@@ -723,12 +723,12 @@ QDomElement SVGExPlug::processTextItem(PageItem *Item, QString trans, QString fi
 						if (charStyle.underlineWidth() != -1)
 							lw = (charStyle.underlineWidth() / 1000.0) * (charStyle.fontSize() / 10.0);
 						else
-							lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+							lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 					}
 					else
 					{
 						Upos = charStyle.font().underlinePos(charStyle.fontSize() / 10.0);
-						lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+						lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 					}
 					if (charStyle.baselineOffset() != 0)
 						Upos += (charStyle.fontSize() / 10.0) * (charStyle.baselineOffset() / 1000.0);
@@ -780,12 +780,12 @@ QDomElement SVGExPlug::processTextItem(PageItem *Item, QString trans, QString fi
 						if (charStyle.strikethruWidth() != -1)
 							lw = (charStyle.strikethruWidth() / 1000.0) * (charStyle.fontSize() / 10.0);
 						else
-							lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+							lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 					}
 					else
 					{
 						Upos = charStyle.font().strikeoutPos(charStyle.fontSize() / 10.0);
-						lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+						lw = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 					}
 					if (charStyle.baselineOffset() != 0)
 						Upos += (charStyle.fontSize() / 10.0) * hl->glyph.scaleV * (charStyle.baselineOffset() / 1000.0);
@@ -970,12 +970,12 @@ QDomElement SVGExPlug::processPathTextItem(PageItem *Item, QString trans, QStrin
 					if (charStyle.underlineWidth() != -1)
 						Uwid = (charStyle.underlineWidth() / 1000.0) * (charStyle.fontSize() / 10.0);
 					else
-						Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+						Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 				}
 				else
 				{
 					Upos = charStyle.font().underlinePos(charStyle.fontSize() / 10.0);
-					Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+					Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 				}
 				if (charStyle.baselineOffset() != 0)
 					Upos += (charStyle.fontSize() / 10.0) * (charStyle.baselineOffset() / 1000.0);
@@ -1046,12 +1046,12 @@ QDomElement SVGExPlug::processPathTextItem(PageItem *Item, QString trans, QStrin
 					if (charStyle.strikethruWidth() != -1)
 						Uwid = (charStyle.strikethruWidth() / 1000.0) * (charStyle.fontSize() / 10.0);
 					else
-						Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+						Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 				}
 				else
 				{
 					Upos = charStyle.font().strikeoutPos(charStyle.fontSize() / 10.0);
-					Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), 1.0);
+					Uwid = qMax(charStyle.font().strokeWidth(charStyle.fontSize() / 10.0), qreal(1.0));
 				}
 				if (charStyle.baselineOffset() != 0)
 					Upos += (charStyle.fontSize() / 10.0) * (charStyle.baselineOffset() / 1000.0);
